server/python-dlib/requirements.txt
face_recognition
python-socketio
eventlet
flask
aiohttp